Update LICENSE
The text you provided is the MIT License. This license is a permissive open-source license that grants users the freedom to use, modify, distribute, and sell the software under certain conditions. Here is a breakdown of the license terms:

The license grants permission to any person to use, copy, modify, merge, publish, distribute, sublicense, and sell copies of the software and associated documentation files free of charge.
The license requires that the above copyright notice and permission notice be included in all copies or substantial portions of the software.
The software is provided "as is" without any warranty, including but not limited to warranties of merchantability or fitness for a particular purpose.
The authors or copyright holders of the software shall not be liable for any claims, damages, or liabilities arising from the use or distribution of the software.
In summary, the MIT License allows you to use, modify, and distribute the software for any purpose, provided that you include the original copyright notice and disclaimer of warranty in all copies or substantial portions of the software.