Create config.py
The code you provided seems to be a configuration file for a project. It sets several variables that control the behavior and settings of the project. Here's a breakdown of the configuration:

VOCAB_SIZE: This variable represents the size of the vocabulary, typically used in natural language processing tasks. It determines the number of unique words or tokens that will be considered in the project.

USE_CUDA: This boolean variable indicates whether the project will utilize CUDA (Compute Unified Device Architecture) for GPU acceleration. If set to True, it suggests that the project is designed to leverage GPU resources for faster computation.

DEVICES: This list variable contains the indices or IDs of the CUDA devices to be used in the project. It specifies which GPUs will be utilized when USE_CUDA is enabled.

CUDA_DEVICE: This variable represents the index or ID of the CUDA device to be used as the primary device. It typically corresponds to the first device in the DEVICES list.

VERSION: This variable denotes the version number of the project. It helps track the evolution and updates of the project over time.

MAXLEN: This variable specifies the maximum length or size of a sequence or input in the project. It can be used to limit the length of sentences, documents, or any other input data.

LEARNING_RATE: This variable determines the learning rate for the project's optimization algorithm. It controls the step size or rate at which the model's parameters are updated during training.

These variables can be adjusted and customized according to the specific requirements and goals of the project.