Create vocab.py
The code you provided is a Python implementation of a vocabulary builder for natural language processing tasks. It defines two classes: TorchVocab and Vocab.

The TorchVocab class represents a vocabulary object used to numericalize text data. It has attributes such as freqs (a counter object holding the frequencies of tokens), stoi (a dictionary mapping token strings to numerical identifiers), and itos (a list of token strings indexed by their numerical identifiers). The class provides methods for initializing the vocabulary, loading pretrained word vectors, and extending the vocabulary with additional words.

The Vocab class is a subclass of TorchVocab and adds functionality specific to text processing tasks. It sets predefined special tokens (such as <pad>, <unk>, <eos>, <sos>, and <mask>) and assigns their corresponding numerical indices. The class also includes methods for converting sentences to numerical sequences and vice versa, as well as loading and saving the vocabulary object.

The WordVocab class is a subclass of Vocab and specializes in building vocabularies from text data. It counts the frequencies of words in the provided text and constructs the vocabulary accordingly. The class includes methods for converting sentences to numerical sequences and vice versa, based on the vocabulary.

The build() function is a utility function that can be used to build a vocabulary from a corpus. It takes command-line arguments specifying the corpus file path, output file path for the vocabulary, vocabulary size, encoding, and minimum frequency of words to include in the vocabulary. It reads the corpus file, constructs the vocabulary using the WordVocab class, and saves the resulting vocabulary object to the specified output file.

To use this code, you can import the necessary classes and functions and then call the build() function with the desired arguments. For example:

python

from your_module import build

# Call the build() function with the desired arguments
build(corpus_path="corpus.txt", output_path="vocab.pkl", vocab_size=10000, min_freq=2)
This will read the text data from the "corpus.txt" file, build a vocabulary with a maximum size of 10000 tokens, and a minimum frequency of 2. The resulting vocabulary will be saved as a pickle file named "vocab.pkl".