Create how2use.py
The code you provided further demonstrates the usage of the UsableTransformer class to encode text using a pre-trained model. Here's a breakdown of the code:

Import statements:

os: This imports the module for interacting with the operating system.
from config import *: This imports variables and settings from a file named config.
torch.nn: This imports the module for defining neural networks in PyTorch.
scipy.ndimage.filters.gaussian_filter1d: This imports a function from SciPy for applying a 1-D Gaussian filter to an array.
torch.autograd.Variable: This imports the Variable class from the torch.autograd module.
torch: This imports the main PyTorch module.
numpy: This imports the NumPy library.
Importing from other files:

import eval_utils as utils: This imports a module named eval_utils and assigns it the alias utils. This module likely contains utility functions for evaluation.
Creating an instance of UsableTransformer:

palmtree = utils.UsableTransformer(model_path="./palmtree/transformer.ep19", vocab_path="./palmtree/vocab"): This line creates an instance of the UsableTransformer class from the eval_utils module. It loads a pre-trained model from the file transformer.ep19 located in the ./palmtree/ directory and a vocabulary from the file vocab located in the same directory.
Encoding text:

text = ["mov rbp rdi", ...]: This defines a list called text that contains multiple instruction strings.
embeddings = palmtree.encode(text): This line encodes the given text using the encode method of the UsableTransformer instance (palmtree). It returns the embeddings of the text, which are stored in the embeddings variable.
Printing the results:

print("usable embedding of this basicblock:", embeddings): This line prints the embeddings of the text.
print("the shape of output tensor: ", embeddings.shape): This line prints the shape of the embeddings tensor.
Overall, this code demonstrates how to use the UsableTransformer class to encode text using a pre-trained model. The encoded representations can then be used for further analysis or downstream tasks.